Issue #37 Add Get by id for Terms, Weekdays, Slots

+++ SSA2020-Back-Hypnotized-Chicken/SSA2020-Back-Hypnotized-Chicken.DataAccessLayer/Repositories/Departments/DepartmentsRepository.cs
@@ -29,7 +29,7 @@ namespace SSA2020_Back_Hypnotized_Chicken.DataAccessLayer.Repositories.Departmen
         {
             return _dbContext.Departments.ToList();
         }
-        public Department GetDepartmentById(int id)
+        public Department GetDepartmentById(short id)
         {
             return _dbContext.Departments.FirstOrDefault(d => d.Id == id);
         }
